Added support for single-click open
Nautilus has support for opening files with a single click, instead of a double-click,
by setting the dconf key "org.gnome.nautilus.preferences.click-policy" to "single".
This patch reads and tracks the value of that key and honors it to keep Desktop icons
behavior consistent with Nautilus.

Avoid opening the file when moving it with "single-click" mode enabled

The previous patch didn't allow to move the icons because it opened the file with
the press event. Changed to the release event, thus allowing to click and move the
icon into a new possition. It also doesn't open twice a file if the user double
clicks on it when the "single-click" mode is enabled.

Avoid opening an icon when a multiple selection ends over an icon

If the "single click" mode is used, starting a lazo selection and releasing the
left button over an icon also launched the icon. This patch fixes this, launching
the icon only when the user clicks without moving the mouse to start a selection.

Removed unneeded blank space

Used spaces instead of tabs

Another forgotten tab

desktopManager: Split out findMonitorForPos() helper

It makes the code more readable and gives us a common place to add
compatibility with different shell versions.

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/12

desktopManager: Adjust for global.screen removal in 3.30

This should be enough to make the extension work in both 3.28 and
3.30.

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/12

Several fixes asked for merging

Several changes and fixes asked by the author for merging

* Use camelCase style in the new variables and properties
* Move the nautilus settings code into settings.js
* Change "lazo selection" with "rubberband selection"

If one branch in an IF has brackets, it should have them in both branches

Removed tabs

Single line IFs don't use brackets

Adjusted _onChangedNautilusSettings to the standard pattern (_onNautilusSettingsChanged)