Adds support for showing the HOME, Trash and other user folders
Allows to show the Home, Trash, Documents, Downloads, Pictures, Videos and
Music folders in the desktop

Fix to detect when a new file is added to the trash

The creation of a file was not correctly managed in the monitoring. Creating
a file in the desktop is detected fine, but a file moved inside the trash is not
detected. Deleting a file from the desktop changes the trash can icon because the
delete operation is detected in the desktop, and all icons are refreshed. But
without this fix, deleting a file from Nautilus, located in other folder, is not
detected and the trash icon is not updated.