Fixes the bug in the rubberband selection when it ends over an icon
When doing a rubberband selection, if the user releases the button when the mouse
cursor is over an icon, the selection isn't terminated; instead, moving the cursor
without the button pressed still selects more icons, and only pressing and releasing
it ends the selection.

The problem is that the _onReleaseButton callback in the icon class blindy stops the
event propagation when the button release is button 1, without taking into account if
the user really pressed on it or not.