Fix error message when the desktop icons are reloaded
Every time the desktop has to be refreshed (like when a file or folder is deleted
or created), mutter outputs this error (repeated several times, exactly once per
icon in the desktop):

(gnome-shell.bin:1348): Gjs-CRITICAL **: 17:01:05.735: Attempting to call back
into JSAPI during the sweeping phase of GC. This is most likely caused by not
destroying a Clutter actor or Gtk+ widget with ::destroy signals connected, but
can also be caused by using the destroy(), dispose(), or remove() vfuncs. Because
it would crash the application, it has been blocked and the JS callback not invoked.

(gnome-shell.bin:1348): Gjs-CRITICAL **: 17:01:05.735: The offending signal was
destroy on StBin 0x5625dcb5f0c0. == Stack trace for context 0x5625d9dac4b0 ==