Fixes critical errors during bootup
When Gnome-Shell is launched from GDM3 and has enabled the Desktop Icons extension,
several critical errors are shown in the console, like:

  (gnome-shell.bin:6393): Gjs-CRITICAL **: 18:13:23.999: Object St.Widget
  (0x5648ac6b7150), has been already deallocated - impossible to access to it. This
  might be caused by the fact that the object has been destroyed from C code using
  something such as destroy(), dispose(), or remove() vfuncs

  == Stack trace for context 0x5648aa7af320 ==

They are caused because the extension is loaded before the desktop is ready.

The fix checks, during the extension launch, whether the shell startup process
has been completed or not. If it is still in progress (which happens when the
extension is loaded during Gnome Shell startup), it subscribes to the
"startup-complete" signal in order to continue only when the shell is ready.

On the other hand, if the extension was disabled and it is enabled manually by
the user (like when it is installed from the Gnome Extensions website), since
the shell startup process have been already completed, the extension starts
inmediately.

This process ensures that all the elements in Gnome Shell needed by the
desktop-icons extension have been created and are in place and ready.