Align icons to the top
Icons and their text are, currently, being aligned to the middle of the grid.
This means that when the icon name has two lines, the icon is placed a little
bit upper than the ones with one line. This change fixes this, putting all
the icons aligned to the top.