Fix icons jumping after every refresh
By default, all files are painted secuentially, in the order provided by
the call that returns the content of the desktop folder.

Each file at the desktop folder store their current coordinates along them
at "metadata::nautilus-icon-position" property, using the metadata
capabilities from the filesystem. This allows to paint them in the same
place every time the computer boots. But when a file is new (because it
has been created, copied or moved), it must be assigned new coordinates.
These are, by default, (0,0).

When the desktop is repainted, the icons are painted strictly in the same
order returned by the operating system when the list of files in the
desktop folder is required. This means that, if a new file (which still
doesn't have assigned new coordinates) is located in the list before a
file with coordinates (0,0), it will be painted at (0,0); then, when the
algorithm reaches the file that originally was located at (0,0), it finds
that those coordinates are already in use, so it will try to relocate
that file to the nearest free location. This can result in a chain
reaction, moving several files out of place only by adding a new file
to the desktop.

The solution consists in doing the paint process in two passes: in the
first one, only those files that already have assigned coordinates in
their "metadata::nautilus-icon-position" entry are painted, thus ensuring
that they will be painted in the right position; and in the second pass,
only the files that still doesn't have assigned coordinates. This ensures
that the new files will fill the free places, instead of pushing other
files out of their original place.

Co-author: csoriano