Removed unnecesary blank spaces and relocated braces
The code had several lines containing only blank spaces, and also
some lines ended with a blank space before the carriage return. This
patch removes those unnecesary blank spaces.

Also it moves the open braces in functions to the same line in some
functions that have them in the next line, to ensure consistency
in the coding style.