Removed sqrt in equation for distance between two points
When relocating an icon in another place because its original place is
already occupied by another icon, or when searching an empty place for
a new file, distance from the original place (or 0,0 when it is a new
icon) is used to find the nearest free position. Since the true distance
is not needed, but only compare between distances, it is possible to
remove the square root in the ecuation, because if sqrt(a) < sqrt(b),
then a < b.