Changed grid parameters from an array to a dictionary
The grid parameters (grid width and height, and icon size) are stored in
a simple array, which gives poor information about what means each value.

This patch replaces the array with an associative array or dictionary,
where each value is addresed by an string that specify the setting that
corresponds to it, instead of using its possition.