Add a separation around each selected icon
Now, with the selection width always set to the maximum, when several
adjacent icons are selected the result is an uniform "blue block" with
the icons inside. This change adds a separation between icons, helping
to keep the visual notion of several elements selected, instead of a
single big one.

It is a must to put the pading in the file-item part instead of the
file-item:selected because doing the later makes the icon move downwards
each time it is selected.

It also adds a 1px border to visually enhance it.