Optimize finding the best location for a new icon
The code used to find the best location for a new icon when they are added
to the desktop performs a sweep over all the grid elements for each icon.
This is very inefficient because the majority of the icons are going to be
put in their original location, and only a very small number, if any, will
require finding the nearest free place.

This patch transforms the original icon coordinates into the final row and
column, and checks if there is already an icon. If not, that position will
be assigned without performing the sweep.

This algorithm reduces the time needed to repaint 30 icons from about 180
milliseconds to approximately 30 ms.