Fix error when connecting an external monitor
When having a laptop and connecting an external screen, the desktop is
rebuilt; unfortunately there is this error message:

Object St.Widget (0x5557c3a01a70), has been already deallocated -
impossible to access to it. This might be caused by the fact that the
object has been destroyed from C code using something such as destroy(),
dispose(), or remove() vfuncs
== Stack trace for context 0x5557c1bcb4d0 ==

This fix connects to the destroy signal of the actors in the grid objects
and remove a grid from the grid list when its actor has been destroyed.