Fix icon size problems including hiDPI problem
The icons plus their file name has some problems. Depending on the screen
resolution, they can fit fine, or lack too much space. It also lacks
scale and hiDPI support. The problem seems to be that the size specified
for each element (icon plus text) is used as a maximum value, instead of
the minimum needed to ensure that both fit.

This patch fixes this problem by calculating a number of rows and columns
that guarantee that each grid element is, at least, as big as the desired
size, thus ensuring that each element in the icon grid will have always
enough space. It also adds support for hiDPI screens using this mode.