Monitorize the trash icon from the fileItem
Previously, sending a file to the trash or emptying it triggered a desktop
repaint, which results in the icons flickering, when the only change was
the trash icon.

This patch moves the trash folder monitorization inside the fileItem that
represents the trash, allowing to keep the icon updated without having to
repaint all the desktop icons.