Fix launching .desktop files with field codes
.desktop files can add field codes to the EXEC field to specify extra data
needed to launch a program. For example, %F will be replaced by a list of
files separated by blank spaces, that will be processed as a command line
parameter by the program being launched. Unfortunately, the current code
in desktop-icons extract that line as-is and uses it to launch the
program without doing any replacement, so .desktop files that include
field codes fail to launch.

This patch fixes this by creating a Gio.DesktopAppInfo object with the
.desktop file, and using it to launch the program itself, and also
takes advantage of its methods to extract the icon path.