Refresh icons instead of the whole desktop
Currently, every time a file is added, deleted, renamed... in the desktop,
the whole grid is destroyed and the desktop content is reloaded and
recreated from scratch, which is time-consuming and also produces an
annoying flicker.

This patch uses a monitor to keep track of the files in the desktop,
modifying the icons on-the-fly instead of recreating everything.