Add SHIFT support for rubberband selection
Currently, when using rubberband in the desktop, it is not possible to
join several selections, because every time the left button is released
and pressed again to do a second selection, the first one is erased.

This patch allows to join several rubberband selections: if the SHIFT
key is pressed when the user presses the left button in the mouse, the
current rubberband selection won't be erased, and the new selection will
be added to the previous one.