Allow both SHIFT and CONTROL keys for multiple selection
Currently, to select several files the user must press SHIFT key. But
the SHIFT key is used in Nautilus to do a range selection, this is, to
select all the files located between the first selected file up to the
last one, while the key used to do individual selections is CONTROL. This
makes the desktop work in an inconsisten way.

This patch allows to do multiple selection any of both keys, and also to
allow to join several rubberband selections using also any of SHIFT or
CONTROL keys.