Creates the thumbnails when they aren't available
It is not possible to generate the thumbnails from the extension itself
because some of them can need too much time, and the thumbnail functions
doesn't have an asynchronous version. Doing it can result in too much
time with the main loop blocked while waiting for the thumbnail generator
to end. Also, thread functions aren't available in javascript.

This patch resolves this by creating the thumbnail in a different script
launched in a separate process, and updating the icon when the thumbnail
is available.