Add symlink emblem to symbolic links
When a file in the desktop is a symbolic link, it should show the "symlink
emblem", the little curved arrow in the lower-right corner specifying that
the file isn't a normal file, but a symbolic link to another file. But
currently the extension lacks it.

This patch fixes this, and adds the emblem to the symbolic links.