Fix rubberband selection stopped when touching top bar
When the user starts a rubberband selection and, while doing it, exists the
desktop (for example by going so high that the mouse cursor enters the top
bar, or if it enters another window in the desktop), when the mouse enters
again in the desktop, the rubberband selection is terminated, no matter
the user still have the mouse button pressed and wants to continue
selecting more files. This is because the current code, as soon as the mouse
cursor leaves the grid actor to enter an actor that doesn't belong to it,
ends the rubberband selection.

This patch modifies this by grabbing the stage when the rubberband starts,
and only terminates it when the user releases the button, or when the mouse
leaves the current screen (because it doesn't support multimonitor yet).

This patch fixes #33