Fix return value in timer callback
The GLib documentation specifies that a timer callback must return FALSE
to cancel the timer and avoid being called again, but the _refreshDesktops()
function returns nothing. Although it means that the caller receives
UNDEFINED as the return value, and it should evaluate to FALSE, it is better
to return explicitly the correct value.

This patch explicitly returns FALSE in that function to follow the GLib
documentation.