Create thumbnails sequentially
When there is a file that requires its thumbnails to be created for the
desktop, its fileItem object launches a new process, which runs the
"createthumbnail.js" script without blocking the main loop. But when there
are several thumbnails to be created, each process is launched at the same
time. This can result in a load spike.

This patch fixes this by launching the processes sequentially, thus waiting
until one process has ended before launching the next process. This ensures
that there will be a limited number of process being launched and running
simultaneously.