desktopManager: Fix DnD for multimonitor
The code was failing to calculate properly the original grid that
contained the items that were going to be drop.

As a result, the DnD operation was unsuscesfull for multimonitor setups.

Additionally, the code was more complex that need to be, in part because
of the special situation where we need to remove the item from the
original grid and later on calculate the new coordinates based on its
current position, which was breaking for being removed already.

To fix it, the code performs three pases:
- First, removes all items from the grid, while saving the coordinates
in a temporary data structure
- Second, assign the new coordinates based on the original coordinates
and the drag path lenght.
- Third, performs the drop into the proper grids, at once for each grid
so the code calculating drop positions behaves smoother.