Fix icons being put always in the first screen
When the user tries to move an icon into any screen different than the
first one, it is always moved to the first one. The problem is because,
when calculating the final coordinates of the DnD, they were being
transformed to the coordinates relative to the origin of the destination
actor.

This patch fixes it by passing as the final coordinates the untransformed
coordinates from the release-button event.