Fix not being able to move icons in secondary screens
When the user has several screens and tries to move an icon from any
screen different than the main one, it doesn't work. This happens
because, when sending the coordinates where to drop the icon, they
are previously converted to values relative to the destination actor,
instead of sending them as-is, in absolute values. This works in the
main screen because, in it, relative and absolute coordinates have
exactly the same value, but that isn't true for all the other screens.

This patch fixes this by removing the transformation from absolute into
relative coordinates, thus allowing to move the icons in all the screens
(but only inside the same screen, this patch still doesn't allow to move
them between screens).