Rework of the fileItem remove code in Grids
When a Grid is destroyed it lets its internal widget to also destroy the
icons. The problem with this approach is that those icons are part of a
fileItem, an object shared with desktopManager; thus destroying a Grid
means that those fileItems will remain in an inconsistent state.

This patch ensures that destroying a Grid will remove the icons before,
thus preserving them and allowing to be reused, if needed, after destroying
the Grid. It also improves the code by removing the need of searching in all
the containers an specific fileItem when it has to be removed.