Fix using icon coordinates before being allocated
When a new file is put in the desktop, it doesn't have saved coordinates,
thus when adding this into a grid, it takes the coordinates of its icon as
the new saved coordinates. Unfortunately, when the code gets the icon
coordinates, it still hasn't been allocated by Clutter, so they are invalid.

To fix this, this patch uses the placeholder coordinates. These are valid
because the extension waits until the DesktopGrid object and all its
internal widgets have been allocated before adding inside the fileItems and
their icons.