general: Rename files instead of refreshing desktop
When a file was renamed, the extension forced a full desktop rebuild, which
results in flicker and CPU usage.

This patch modifies both desktopManager.js and fileItem.js to manage directly
the RENAME events from desktopManager and ask the corresponding fileItem to
update its name on-the-fly. Also, this patch sets the basis for future direct
event management, like adding or deleting files on-the-fly.