Refactor _createEmblemIcon
The _createEmblemIcon function in the fileItem.js file creates and returns
a Gio.EmblemedIcon, either from an stock or systemwide-available icon, or
from a picture file. But to be able to use it, it must be put inside an
St.Icon before being inserted in the container of a fileItem.

This patch changes this function, so now it returns directly an St.Icon,
instead of merely returning the icon and forcing the caller to do the job.
This change makes the code cleaner and simpler.