Fix the observer pattern
In the original code, the desktopGrid emits a "selected" signal in name of
the fileItems, instead of asking them to do it. This breaks the observer
pattern, because the signal can be emitted without being known by the
fileItem.

This patch fixes it by passing the rubberband limits to each fileItem and
letting them to decide whether they are selected or not.