general: Introduce concept of file state
Since file operations are inherently slow, it might take some time
to reflect in the UI a change the user has done.

However, not doing anything until it really happens would confuse the
user. The UI needs to reflect the intention of the user, rather than the
actual system layout.

In Nautilus we have the concept of "gone" and other methods for this.
Let's implement here something more dynamic, by providing states.

Rigth not it's not used, but we will use it in later commits to
implement drag and drop and its move operation. In the future this
will be extended for regular copy/paste operations too, and other
possible slow operations.