desktopManager: Implement DnD for items
This is a initial implementation of DnD for items in the desktop.

There is some crudeness since we cannot know the result of the operation
and therefore marking a file as not consistent in the UI might make it
dissapear withouth further signal.

Apart from that, there is not feedback when hovering over droppable
targets.

All of this will come in future works, specially once we have a better
system to track changes.