Fix the desktop not adding or removing icons
When a new file is created in the desktop, or a file is deleted, the desktop
is repainted but the file list isn't refreshed, so new files aren't added
and old files aren't removed.

This patch fixed this by calling scanFiles() instead of
scheduleReLayoutChildren() when a file at the desktop is added, deleted,
moved or renamed.