Show the display name of launchable desktop files
When displaying a desktop file that we'll launch (executable), we
currently show the filename. But desktop files come with a name - let's
show that. We don't want to reopen CVE-2017-14604 though, so only do
this for files with `metadata::trusted' set to 'true'. Future work would
be to consider a UI for this, or an alternative solution.