general: Hide the hidden files when needed
Hidden files are those that start with a dot. These files shouldn't
be shown if the key 'show-hidden' from '/org/gtk/settings/file-chooser'
if 'false'. Unfortunately, the current code doesn't honor this flag and
shows all the files in the desktop, no matter whether they are hidden
or not.

This patch fixes this by checking that flag and the first character in
the visible file name to decide whether a file must be hide or not. It
also connects to the 'changed' signal to ensure that a change in the
settings flag will be also honored.

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/46