Watch for the execute bit changing on .desktop files
We should only ever show the name of the .desktop file if it is
executable. That is:

  - metadata::trusted true
  - executable

With this commit, we use a file monitor to watch for the attributes
changing, and refresh the icon when they do. If the file goes +x or -x,
we will update the icon accordingly (hiding or showing the display name
from the desktop file).