build: Install files in the proper systemwide location
When installing in the system, we need to put the extension in
${prefix}/share/gnome-shell/extensions/<uuid>. Up until now we were just
installing in ${prefix}.

The schema compilation needs tweaking a bit, because we now install
schemas in the proper location under the prefix and we shouldn't do this
when building a package (DESTDIR in the environment, approach copied
from Recipes). We do still need to have the export-zip.sh script call
glib-compile-schemas for the zips we'll upload to extensions.gnome.org
though - when the extension is loaded in this way it uses schemas from
schemas/ within its own tree and nothing will compile them for us there.