desktopManager: Fix dragging folder into itself error
When the user starts a DnD operation, and the destination folder is in
the list of selected items to drag, the operation fails and gnome
shell becomes unstable.

This patch fixes this by removing any appearance of the destination
folder in the list of the files/folders to move, and calling
recreateDesktopIcons() instead of scheduleLayoutChildren().

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/48