desktopManager: fix DragNDrop lack of precission
The DragAndDrop operation currently uses the stored coordinates as the
initial coordinates, and adds to them the displacement made with the
mouse. This is incorrect for two reasons:

- The stored coordinates can be located in any point of the grid
element, so some displacements can result in the icon being located in
a grid position adjacent to the logical/desired one. An example: the
coordinates are located near the right corner of the grid, the user
starts the DnD operation near the left corner of the grid, and ends it
near the right corner of another grid element. In this case, the icon
can be placed in the grid location at the right of the desired one
(the grid location under the mouse).

- If the user switches between screens with different resolutions
(like when using a laptop with an external monitor) it can happen
that an icon has stored coordinates outside the current screen. In
this case the icon is shown in the nearest grid location, but the
stored coordinates aren't overwritten to ensure that, when
connecting the bigger screen, the icon will be shown in the original
coordinates (this is: to avoid that connecting an smaller screen will
change the icon positions). But if the user tries to move this icon,
since the code uses the stored coordinates instead the real ones, it
will end in an incorrect grid location.

This patch fixes this by using the stored coordinates only to locate
the icon when it is put in the desktop. For all the other operations,
only the current real coordinates are used.

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/57