general: fix grab focus
The grabbing of the focus in the desktop doesn't work, so hotkeys also
fail. This is due to the grabHelper, that prevents the
grab_key_focus() method to work.

This patch forces grabbing the focus after releasing the grabHelper,
ensuring that, when the user clicks on the desktop or selects a file,
the focus is grabbed by the desktop. The focus is always set to the
desktop 0 because it doesn't matter which one of all the desktops
receive it, since all the key actions are done over the list of files
stored in this._selection.