fileItem.js: fix undefined property error
When a new file is added to the desktop, it doesn't have the
"nautilus-icon-position" metadata entry. Unfortunately, in this case
the "this._savedCoordinates" is not defined, so when the code tries to
paint it, it fails.

This patch fixes this by defining it with NULL value when there are no
stored coordinates in the metadata.