prefs: Handle translations when installed system-wide
The current code assumes that the extension is installed in the user's
home, while the build system does support system-wide installation.
Account for that by checking for the presence of a locale/ subfolder
in the extension directory, and fall back to gnome-shell's own locale
directory otherwise (assuming a common prefix between extension and
gnome-shell).

https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/66