desktopGrid: Constrain grid to workarea
The main actor is a stack widget, meaning, we use it so we can have
a stack of the grid with icons with floating actors such as the rename
popover.

To achieve that, the FixedLayout of clutter is used. However, that
layout doesn't allocate children based on its size or aligment. So to
perform proper layouting we need to constraint the actual grid size
and not the fixed layout actor.

What happened as consequence is that the icons had the wrong
coordinates asigned, since allocation of the main actor is not dependent
of its descendents.

To fix it, we apply the constraint of the workarea to the grid directly.

Thanks Sergio for the investigation and partial implementation for this!