desktopGrid.js: Remove 'reserved' code in DnD
The method _getEmptyPlaceClosestTo() can receive a dictionary with
reserved positions. This dictionary contains row,column coordinate
pairs of places that must not be used to place a new fileItem in
the grid. But this is redundant because a pair or row,column
coordinates can be in that 'reserved dictionary' only if there is also
an actor in the grid in those coordinates. And since the
_getEmptyPlaceClosestTo() already ensures that the returned coordinates
will always be to a grid element where there is no actor, the result
is that it that this piece of code is redundant and unnecesary, and
that's the reason for this patch for removing it.