desktopManager.js: do nothing if drag folder in itself
When the user drags several files and folders inside a folder, if the
destination folder is in the group it is silently removed, but the
other files and folders are moved inside.

This patch changes this behavior, cancelling the drag'n'drop operation
if the destination folder is between the moved group of files.

Fixes https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/79