desktopManager.js: Ensure end rubberband selection
With DashToDock enabled, if the user starts a rubberband selection and
accidentally ends it over the dock, the rubberband remains active. But
pressing the mouse button in the desktop, although ends the selection,
leaves the extension in an invalid state: not only it doesn't release
naturally the focus (only when the user presses the left button in a
window), but also the rubberband selection is still active in a
"ghost mode".

This patch fixes it by ending the rubberband selection if it is active
and motion is detected withouth the mouse left button pressed.

Fixes https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/83