desktopGrid: fix DnD in multimonitor systems
The Drag'N'Drop operations doesn't work as expected when the user has
a multimonitor environment. This is because the DnD operations expect
to receive the cursor coordinates relative to the receiving widget.
When the user just moves one or several icons, the Drop operation
is sent to a Grid object, which calculates, from the local coordinates,
the global ones in order to proceed to move the icons to the desired
destination. Unfortunately, this calculation was made using the grid
coordinates, but the widget that receives the Drop operation is, in
fact, "this.actor", an St.Widget that contains the grid itself.

To fix this, this patch uses the "this.actor" coordinates to calculate
the absolute drop coordinates.

Fix https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/69