desktopManager: ensure extra folders are inserted first
Sometimes, due to the coordinates problem, there are files and/or
folders that have the same coordinates than the trash or the home
folder. In this case, it is better to give priority to the extra
files, thus ensuring that only the other files are moved, and not
those two. This patch ensures that.