desktopGrid: fix multimonitor DnD inside folders
Until now, DnD inside folders only works when done in the main
monitor. The reason seems to be that the icon grids are actors
inside another actors, but only the grids have the MonitorConstraint
set. Due to this, the container actors are, in fact, "layered" all
in the main monitor, while the grids "inside" are each one in its
corresponding monitor. And, unfortunately, the container actor is
needed to allow to show the Rename popup.

Also, just setting the MonitorConstraint in the container actor isn't
enough, because in that case the inner grid actors doesn't expand
until the icons have been added, which is a problem in the code (check
https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/merge_requests/85 )

This patch fixes this by setting the MonitorConstraint in the
container actor, thus moving all the container actors to their
corresponding monitor and with the right size, and also adding a
BinLayout to each container actor, thus ensuring that the inner
grid will have their "x_expand" and "y_expand" properties honored.