README: fix locale building
When manually compiling the repository and installing the extension
using meson/ninja, the locales aren't installed in the extension
folder but at $HOME/.local/share/locale. This patch adds a note in
the README file to specify how to fix this.

Fix https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/77