general: Add support for "Open with" command in file menu
Currently, the only option for opening an icon is to left-click on
it, or right-click and choose "Open". This forces the user to open
desktop elements only with the default application.

This patch adds a new entry in the right-click menu, "Open with...",
that allows to choose which application launch to open an element.
This is done by using the Freedesktop.org Portal DBus interface.

Fixes https://gitlab.gnome.org/World/ShellExtensions/desktop-icons/issues/32