general: unify parameter names in async callbacks
The asynchronous functions pass to the callback two parameters: the
original object over which the async call was made, and the operation
result. In the code, the names for both parameters varied, using
'source', 'o' and 'obj' for the first one, and 'res' and 'result' for
the second one.

This patch unifies the naming to 'source' and 'result' in all async
method calls.