general: Allow "Open with" only in single selections
The Portal interface only allows to select an application for a single
file. This means that, when several files are selected, it is not
possible to choose a single application for opening all of them.

In the previous MR, the "Open with..." option allowed to open the file
in which the user did the right click to show the menu, and it didn't
matter if there were more files selected. Since this can be confusing,
this patch changes this, and now the "Open with..." option is
disabled when there are more than one file selected.